@@ -3,42 +3,21 @@
   "version": "1.0.0",
   "private": true,
   "scripts": {
-    "build": "rm -rf ./dist; sass src/css:dist/css --style=compressed --load-path=src/css --load-path=node_modules; vue-cli-service build --no-clean",
     "watch": "sass src/css:dist/css --watch --load-path=src/css --load-path=node_modules",
     "watch:dv": "sass src/css/dv-custom.scss:dist/css/dv-custom.css --watch --load-path=src/css --load-path=node_modules",
-    "serve": "vue-cli-service serve",
-    "lint": "vue-cli-service lint"
+    "dev": "vite",
+    "build": "rm -rf ./dist; sass src/css:dist/css --style=compressed --load-path=src/css --load-path=node_modules --no-source-map; vite build",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "core-js": "^3.6.4",
-    "vue": "^2.6.11"
+    "vue": "^3.2.30"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.3.0",
-    "@vue/cli-plugin-eslint": "~4.3.0",
-    "@vue/cli-service": "~4.3.0",
-    "babel-eslint": "^10.1.0",
-    "bootstrap": "^4.4.1",
-    "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^6.2.2",
-    "sass": "^1.38.2",
-    "html-webpack-inline-source-plugin": "1.0.0-beta.2",
-    "html-webpack-plugin": "^4.3.0",
-    "vue-svg-loader": "^0.16.0",
-    "vue-template-compiler": "^2.6.11"
-  },
-  "eslintConfig": {
-    "root": true,
-    "env": {
-      "node": true
-    },
-    "extends": [
-      "plugin:vue/essential",
-      "eslint:recommended"
-    ],
-    "parserOptions": {
-      "parser": "babel-eslint"
-    },
-    "rules": {}
+    "@vitejs/plugin-vue": "^2.2.0",
+    "bootstrap": "^4.6.1",
+    "sass": "^1.49.7",
+    "vite": "^2.8.0",
+    "vite-plugin-singlefile": "^0.6.3",
+    "vite-svg-loader": "^3.1.2"
   }
 }
